requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-core
langchain-chroma
langchain-ollama
pydantic
pytest
httpx
jinja2
langchain-text-splitters
redis
python-dotenv
pyodbc
keybert
stop-words
langgraph
langchain-openai
openai